refactor: reordering catalog files
