+++ src/query/catalog/src/catalog/interface.rs
@@ -11,11 +11,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 use std::any::Any;
 use std::sync::Arc;
 
-use common_base::base::Singleton;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_app::schema::CountTablesReply;
@@ -48,78 +46,12 @@ use common_meta_app::schema::UpsertTableCopiedFileReq;
 use common_meta_app::schema::UpsertTableOptionReply;
 use common_meta_app::schema::UpsertTableOptionReq;
 use common_meta_types::MetaId;
-use dashmap::mapref::entry::Entry;
-use dashmap::DashMap;
 use dyn_clone::DynClone;
-use once_cell::sync::OnceCell;
 
 use crate::database::Database;
 use crate::table::Table;
 use crate::table_args::TableArgs;
 use crate::table_function::TableFunction;
-
-pub const CATALOG_DEFAULT: &str = "default";
-
-static CATALOG_MANAGER: OnceCell<Singleton<Arc<CatalogManager>>> = OnceCell::new();
-
-pub struct CatalogManager {
-    pub catalogs: DashMap<String, Arc<dyn Catalog>>,
-}
-
-impl CatalogManager {
-    pub fn get_catalog(&self, catalog_name: &str) -> Result<Arc<dyn Catalog>> {
-        self.catalogs
-            .get(catalog_name)
-            .as_deref()
-            .cloned()
-            .ok_or_else(|| ErrorCode::BadArguments(format!("no such catalog {}", catalog_name)))
-    }
-
-    pub fn instance() -> Arc<CatalogManager> {
-        match CATALOG_MANAGER.get() {
-            None => panic!("CatalogManager is not init"),
-            Some(catalog_manager) => catalog_manager.get(),
-        }
-    }
-
-    pub fn set_instance(manager: Singleton<Arc<CatalogManager>>) {
-        CATALOG_MANAGER.set(manager).ok();
-    }
-
-    pub fn insert_catalog(
-        &self,
-        catalog_name: &str,
-        catalog: Arc<dyn Catalog>,
-        if_not_exists: bool,
-    ) -> Result<()> {
-        // NOTE:
-        //
-        // Concurrent write may happen here, should be carefully dealt with.
-        // The problem occurs when the entry is vacant:
-        //
-        // Using `DashMap::entry` can occupy the write lock on the entry,
-        // ensuring a safe concurrent writing.
-        //
-        // If getting with `DashMap::get_mut`, it will unlock the entry and return `None` directly.
-        // This makes a safe concurrent write hard to implement.
-        match self.catalogs.entry(catalog_name.to_string()) {
-            Entry::Occupied(_) => {
-                if if_not_exists {
-                    Ok(())
-                } else {
-                    Err(ErrorCode::CatalogAlreadyExists(format!(
-                        "Catalog {} already exists",
-                        catalog_name
-                    )))
-                }
-            }
-            Entry::Vacant(v) => {
-                v.insert(catalog);
-                Ok(())
-            }
-        }
-    }
-}
 
 #[derive(Default, Clone)]
 pub struct StorageDescription {
